refactor: Nest costs tools under src/tools/costs/.
Move top-level costs MCP tools (and tests) into a resource folder per writing-mcp-tools conventions.
