+++ src/tools/costs/query-costs.ts
@@ -1,8 +1,8 @@
 import z from "zod";
-import dateValidator from "../utils/dateValidator";
-import paginationData from "../utils/paginationData";
-import MCPUserError from "./structure/MCPUserError";
-import registerTool from "./structure/registerTool";
+import dateValidator from "../../utils/dateValidator";
+import paginationData from "../../utils/paginationData";
+import MCPUserError from "../structure/MCPUserError";
+import registerTool from "../structure/registerTool";
 
 const description = `
 Query for costs in a Vantage Account. These are independent of a cost reports.
