+++ src/tools/index.ts
@@ -6,6 +6,7 @@ import "./business-metrics";
 import "./canvases";
 import "./cost-alerts";
 import "./cost-reports";
+import "./costs";
 import "./create-folder";
 import "./create-report-notification";
 import "./create-virtual-tag-config";
@@ -27,15 +28,12 @@ import "./list-audit-logs";
 import "./list-cost-integrations";
 import "./list-cost-providers";
 import "./list-cost-services";
-import "./list-costs";
 import "./list-folders";
 import "./list-provider-resources";
 import "./list-report-notifications";
 import "./list-tag-values";
 import "./list-tags";
-import "./list-unit-costs";
 import "./network-flow-reports";
-import "./query-costs";
 import "./recommendation-views";
 import "./recommendations";
 import "./resource-reports";
