+++ test/tools/costs/list-costs.test.ts
@@ -1,7 +1,7 @@
 import type { GetCostsResponse } from "@vantage-sh/vantage-client";
 import { expect } from "vitest";
-import tool from "../../src/tools/list-costs";
-import { DEFAULT_LIMIT } from "../../src/tools/structure/constants";
+import tool from "../../../src/tools/costs/list-costs";
+import { DEFAULT_LIMIT } from "../../../src/tools/structure/constants";
 import {
   dateValidatorPoisoner,
   type ExecutionTestTableItem,
@@ -12,7 +12,7 @@ import {
   requestsInOrder,
   type SchemaTestTableItem,
   testTool,
-} from "../../src/utils/testing";
+} from "../../../src/utils/testing";
 
 type Validators = ExtractValidators<typeof tool>;
 type OutputSchema = ExtractOutputSchema<typeof tool>;
