+++ test/tools/costs/list-unit-costs.test.ts
@@ -1,6 +1,6 @@
 import type { GetUnitCostsResponse } from "@vantage-sh/vantage-client";
 import { expect } from "vitest";
-import tool from "../../src/tools/list-unit-costs";
+import tool from "../../../src/tools/costs/list-unit-costs";
 import {
   dateValidatorPoisoner,
   type ExecutionTestTableItem,
@@ -11,7 +11,7 @@ import {
   requestsInOrder,
   type SchemaTestTableItem,
   testTool,
-} from "../../src/utils/testing";
+} from "../../../src/utils/testing";
 
 type Validators = ExtractValidators<typeof tool>;
 type OutputSchema = ExtractOutputSchema<typeof tool>;
